💥 separate dev.jl into multiple scripts
Hopefully will make things easier in the long term. Probably will not though
